tests/wpt/tests/navigation-api/navigate-event/navigate-destination-getState-reload.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    let navState = { statevar: "state" };
    navigation.onnavigate = t.step_func_done(e => {
      assert_equals(e.navigationType, "reload");
      assert_not_equals(e.destination, null);
      assert_not_equals(e.destination.getState(), undefined);
      assert_equals(e.destination.getState().statevar, "state");
      assert_not_equals(e.destination.getState(), e.destination.getState());
      assert_equals(e.destination.key, "");
      assert_equals(e.destination.id, "");
      assert_equals(e.destination.index, -1);
      // NavigateEvent sourceElement is still in development, so test whether it is available.
      if ("sourceElement" in e) assert_equals(e.sourceElement, null);
      e.intercept();
    });
    navigation.updateCurrentEntry({ state: navState });
    location.reload();
  }, 0);
}, "navigate event destination.getState() on location.reload()");
</script>
