LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-destination-getState-reload-expected.txt


FAIL navigate event destination.getState() on location.reload() assert_equals: expected (object) null but got (undefined) undefined
